AI Tutorial Compatiblity
